requirements.txt
# ── Core Framework ──
fastapi==0.115.6
uvicorn[standard]==0.34.0

# ── Database ──
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite>=0.20.0
alembic==1.14.1

# ── Templating ──
jinja2==3.1.5

# ── Authentication ──
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
authlib>=1.3.0
httpx>=0.27.0

# ── Settings ──
pydantic-settings==2.7.1

# ── Utilities ──
python-dotenv==1.0.1
email-validator==2.2.0
